+++ lib/Doctrine/ODM/PHPCR/Mapping/Driver/AnnotationDriver.php
@@ -19,12 +19,12 @@
 
 namespace Doctrine\ODM\PHPCR\Mapping\Driver;
 
-use Doctrine\Common\Persistence\Mapping\ClassMetadata;
-use Doctrine\Common\Persistence\Mapping\Driver\AnnotationDriver as AbstractAnnotationDriver;
-use Doctrine\Common\Persistence\Mapping\Driver\MappingDriver;
 use Doctrine\ODM\PHPCR\Event;
 use Doctrine\ODM\PHPCR\Mapping\Annotations as ODM;
 use Doctrine\ODM\PHPCR\Mapping\MappingException;
+use Doctrine\Persistence\Mapping\ClassMetadata;
+use Doctrine\Persistence\Mapping\Driver\AnnotationDriver as AbstractAnnotationDriver;
+use Doctrine\Persistence\Mapping\Driver\MappingDriver;
 
 /**
  * The AnnotationDriver reads the mapping metadata from docblock annotations.
@@ -229,3 +229,5 @@ class AnnotationDriver extends AbstractAnnotationDriver implements MappingDriver
         return $cascade;
     }
 }
+
+interface_exists(ClassMetadata::class);