+++ lib/Doctrine/ODM/PHPCR/Mapping/Driver/XmlDriver.php
@@ -19,10 +19,10 @@
 
 namespace Doctrine\ODM\PHPCR\Mapping\Driver;
 
-use Doctrine\Common\Persistence\Mapping\ClassMetadata;
-use Doctrine\Common\Persistence\Mapping\Driver\FileDriver;
-use Doctrine\Common\Persistence\Mapping\MappingException as DoctrineMappingException;
 use Doctrine\ODM\PHPCR\Mapping\MappingException;
+use Doctrine\Persistence\Mapping\ClassMetadata;
+use Doctrine\Persistence\Mapping\Driver\FileDriver;
+use Doctrine\Persistence\Mapping\MappingException as DoctrineMappingException;
 use SimpleXmlElement;
 
 /**
@@ -349,3 +349,5 @@ class XmlDriver extends FileDriver
         return $cascade;
     }
 }
+
+interface_exists(ClassMetadata::class);