+++ lib/Doctrine/ODM/PHPCR/Translation/LocaleChooser/LocaleChooser.php
@@ -19,8 +19,8 @@
 
 namespace Doctrine\ODM\PHPCR\Translation\LocaleChooser;
 
-use Doctrine\Common\Persistence\Mapping\ClassMetadata;
 use Doctrine\ODM\PHPCR\Translation\MissingTranslationException;
+use Doctrine\Persistence\Mapping\ClassMetadata;
 
 /**
  * Class to get the list of preferred locales.
@@ -180,3 +180,5 @@ class LocaleChooser implements LocaleChooserInterface
         return $this->defaultLocale;
     }
 }
+
+interface_exists(ClassMetadata::class);